refactor(fleet): extract dashboard HTML to fleet-dashboard-html.ts (P4.1)
The 437-line ACTIVITY_VIEWER_HTML template literal lived at the bottom of
fleet-manager.ts purely as a static asset served by the health server.
Moving it to its own module trims fleet-manager.ts by ~437 lines without
touching any runtime behaviour.

Pure-constant extraction: zero new dependencies, no Context interface
needed, no circular import risk.

P4.1 step 1 of 4.
